Move some provisioning tasks to fab_deploy
