+++ tool/cstar_perf/docker/tasks.py
@@ -3,109 +3,13 @@
 """
 docker.tasks
 
-Various tasks to provision the docker machines.
+Various tasks related to the docker machines provisioning.
 
 """
 
-import textwrap
 from fabric import api as fab
-from fabric.tasks import execute as fab_execute
-from fabric.contrib.files import append as fab_append
-from ilogue.fexpect import expect, expecting, run
+from cstar_perf.tool import fab_deploy
 
-# TODO All python scripts should be python resource file, instead of string.
-
-def run_python_script(script):
-    fab.run("rm -f ~/pyscript.py")
-    fab_append("pyscript.py", textwrap.dedent(script))
-    output = fab.run("python pyscript.py")
-    fab.run("rm ~/pyscript.py")
-    return output
-
-def generate_frontend_credentials():
-    """Create the server keys and application config"""
-
-    # Save the credentials in a file, so we can use it to associate a cluster
-    fab.run("cstar_perf_server --get-credentials > ~/credentials.txt")
-
-def create_default_users():
-    """Create a default admin and normal users"""
-
-    create_users_script = """
-    from cstar_perf.frontend.server.model import Model
-    db = Model()
-    admin = db.create_user('admin@admin.com', 'Admin Full Name', ['user','admin'])
-    db.set_user_passphrase('admin@admin.com', 'admin')
-    user = db.create_user('user@user.com', 'User Full Name', ['user'])
-    db.set_user_passphrase('user@user.com', 'user')
-    """
-    run_python_script(create_users_script)
-
-def add_cluster_to_frontend(cluster_name, num_nodes, public_key):
-    """Add the cluster to the frontend configuration"""
-
-    add_cluster_script = """
-    from cstar_perf.frontend.server.model import Model
-    db = Model()
-    db.add_cluster('{name}', {num_nodes}, '{name}')
-    db.add_pub_key('{name}', 'cluster', '{key}', replace=True)
-    """.format(name=cluster_name, num_nodes=num_nodes, key=public_key)
-    run_python_script(add_cluster_script)
-
-def add_jvm_to_cluster(cluster_name, jvm):
-    """Add a jvm to the frontend cluster configuration"""
-
-    path = "~/fab/jvms/{jvm}".format(jvm=jvm)
-    add_jvm_script = """
-    from cstar_perf.frontend.server.model import Model
-    db = Model()
-    db.add_cluster_jvm('{name}', '{jvm}', '{path}')
-    """.format(name=cluster_name, path=path, jvm=jvm)
-    run_python_script(add_jvm_script)
-
-def get_frontend_credentials():
-    """Get the frontend server keys"""
-
-    # Read the credentials file and return a dict
-    output = fab.run("cat ~/credentials.txt")
-    if 'Server public key' not in output:
-        raise ValueError("credentials.txt doesn't contain proper keys")
-    lines = output.split('\n')
-    public_key = lines[1].split(': ')[1].strip()
-    verify_code = lines[2].split(': ')[1].strip()
-
-    return {'public_key': public_key, 'verify_code': verify_code}
-
-def generate_client_credentials(cluster_name, public_key, verify_code):
-    """Generate the client credentials"""
-
-    prompts = []
-    prompts += expect('Enter a name for this cluster:', cluster_name)
-    prompts += expect("Input the server's public key:", public_key)
-    prompts += expect("Input the server verify code: ", verify_code)
-
-    with expecting(prompts):
-        output = run('cstar_perf_client --get-credentials')
-
-    lines = output.split('\n')
-    client_public_key = [line for line in lines if line.startswith("Your public key is")][0]
-    fab.run("echo '{}' > ~/credentials.txt".format(client_public_key))
-
-def get_client_credentials():
-    """Get the client server key"""
-
-    # Read the credentials file and return a dict
-    output = fab.run("cat ~/credentials.txt")
-    public_key = output.split(': ')[1].strip()
-
-    return {'public_key': public_key}
-
-def get_client_jvms():
-    """Get a list of all jvms of the client"""
-
-    output = fab.run("ls ~/fab/jvms")
-    jvms = [jvm for jvm in output.split(' ') if jvm]
-    return jvms
 
 def setup_client_daemon(node_name):
     """Add or replace cstar_perf_client config to supervisord"""
@@ -127,11 +31,10 @@ def setup_client_daemon(node_name):
     with open(config_path, 'w') as f:
         config.write(f)
     """.format(node_name)
-    run_python_script(setup_script)
+    fab_deploy.run_python_script(setup_script)
     # ensure the client is restarted
-    fab.run("supervisorctl -c /supervisord.conf reread")
-    fab.run("supervisorctl -c /supervisord.conf stop cstar_perf_client")
-    fab.run("supervisorctl -c /supervisord.conf start cstar_perf_client")
+    restart_all_services()
+
 
 def add_or_update_host_ips(hosts):
     """Update /etc/hosts ips"""
@@ -160,7 +63,8 @@ def add_or_update_host_ips(hosts):
         for host, ip in new_hosts:
             f.write("{{}} {{}}\n".format(ip, host))
     """.format(hosts=hosts)
-    run_python_script(update_script)
+    fab_deploy.run_python_script(update_script)
+
 
 def restart_all_services():
     fab.run("test -f /supervisord.conf && supervisorctl -c /supervisord.conf reread")
